+++ src/index.ts
@@ -1,2 +1,3 @@
+export { runWithStateVocab } from "./store"
 export { defineState } from "./state"
 export { setupStorage } from "./setup"